src/etl/db/requirements.txt
psycopg2-binary
python-dotenv
pandas
pyarrow